python-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cassandra-driver==3.28.0
pydantic==2.5.0